i2c-msm-v2: Ensure system is always active for i2c transfers
There are quite a few i2c devices that are intended to remain active while
the system is suspended, which means that when one of them fires off an
interrupt, the subsequent i2c transfers inside the interrupt handler will
fail.

To fix this, make the entire system resume when an i2c transfer is
requested while the system is suspended. i2c transfer requests will be
blocked until the system is completely resumed, so that they won't fail.

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: xNombre <kartapolska@gmail.com>
Signed-off-by: Ahmad Nauval Syahputra <anvalputra@gmail.com>